+++ backend/src/modules/skus/skus.model.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, JoinColumn, ManyToOne } from 'typeorm';
 import BaseModel from '../../../commons/models/base.model';
-import Color from '../colors/colors.model';
+import ProductSize from '../product-sizes/product-sizes.model';
 import ProductColor from '../product-colors/product-colors.model';
 
 @Entity('skus')
@@ -9,9 +9,9 @@ export default class Sku extends BaseModel {
   @JoinColumn({ name: 'product_color_id', referencedColumnName: 'id' })
   productColor: ProductColor;
 
-  @ManyToOne(() => Color)
+  @ManyToOne(() => ProductSize)
   @JoinColumn({ name: 'product_size_id', referencedColumnName: 'id' })
-  productSize: Color;
+  productSize: ProductSize;
 
   @Column({ name: 'price', type: 'numeric' })
   price: number;